Decouple database editor, move tools to their own directory/package
