+++ tools/gacha_simulator.py
