+++ tools/modules/__init__.py
@@ -1,0 +1,1 @@
+from .database_editor import DatabaseEditor
